CloudDatabasePython/requirements.txt
Flask
google-cloud-firestore
python-dotenv
firebase-admin
